Rename MicroApp class to CeleryApp
this change is needed because we are going to add a new FlaskApp to make
the Micro endpoints available through an API Rest
